Moved views into settings.py
