+++ corehq/tabs/tabclasses.py
@@ -1410,7 +1410,7 @@ class ProjectSettingsTab(UITab):
         project_info = []
 
         if user_is_admin:
-            from corehq.apps.domain.views import EditBasicProjectInfoView, EditPrivacySecurityView
+            from corehq.apps.domain.views.settings import EditBasicProjectInfoView, EditPrivacySecurityView
 
             project_info.extend([
                 {
@@ -1423,14 +1423,14 @@ class ProjectSettingsTab(UITab):
                 }
             ])
 
-        from corehq.apps.domain.views import EditMyProjectSettingsView
+        from corehq.apps.domain.views.settings import EditMyProjectSettingsView
         project_info.append({
             'title': _(EditMyProjectSettingsView.page_title),
             'url': reverse(EditMyProjectSettingsView.urlname, args=[self.domain])
         })
 
         if toggles.OPENCLINICA.enabled(self.domain):
-            from corehq.apps.domain.views import EditOpenClinicaSettingsView
+            from corehq.apps.domain.views.settings import EditOpenClinicaSettingsView
             project_info.append({
                 'title': _(EditOpenClinicaSettingsView.page_title),
                 'url': reverse(EditOpenClinicaSettingsView.urlname, args=[self.domain])
@@ -1530,7 +1530,7 @@ class ProjectSettingsTab(UITab):
 
 def _get_administration_section(domain):
     from corehq.apps.domain.views.internal import TransferDomainView
-    from corehq.apps.domain.views import (
+    from corehq.apps.domain.views.settings import (
         FeaturePreviewsView,
         RecoveryMeasuresHistory,
     )
